Create Chapter 3 (Model Validation)
The codes for the validation chapter are similar to the codes for the default model except that for the validation, the model was run using farm data on soil type, rainfall, crop yield and farm areas. As a result, estimation of gross margin and yield penalties were based on the farm yield data from the FBS data. Workable hours were estimated based on farm soil and rainfall data.

The comparison of model generated results were compared with observed data and the statistics measure of association were estimated using the function ValStats (Line 3640).